Add crafting specializations

+++ library/ForSummoners.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.7;
 import "./Rarity.sol";
 
 contract ForSummoners {
-  modifier approvedForSummoner(uint256 summoner) {
+  modifier approvedForSummoner(uint summoner) {
     if (Rarity.isApprovedOrOwnerOfSummoner(summoner)) {
       _;
     } else {
